+++ api/v1alpha1/userconfig/service/pg/pg.go
@@ -477,7 +477,8 @@ type Timescaledb struct {
 }
 type PgUserConfig struct {
 	// +kubebuilder:validation:MaxItems=1
-	// Additional Cloud Regions for Backup Replication
+	// +kubebuilder:deprecatedversion:warning="additional_backup_regions is deprecated"
+	// Deprecated. Additional Cloud Regions for Backup Replication
 	AdditionalBackupRegions []string `groups:"create,update" json:"additional_backup_regions,omitempty"`
 
 	// +kubebuilder:validation:MinLength=8